+++ tensorflow/c/experimental/gradients/gradient_checker.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow/c/eager/gradient_checker.h"
+#include "tensorflow/c/experimental/gradients/gradient_checker.h"
 
 #include <memory>
